web/requirements.txt
bert-serving-client==1.10.0
elasticsearch==8.6.1
Flask==2.2.2